place project archiving into separate controller
