+++ app/services/projects/archive_service.rb
@@ -31,6 +31,7 @@
 module Projects
   class ArchiveService < ::BaseServices::BaseContracted
     include Contracted
+    include Projects::Concerns::UpdateDemoData
 
     def initialize(user:, model:, contract_class: Projects::ArchiveContract)
       super(user: user, contract_class: contract_class)
